manifest.json: Add `author` and `homepage_url`
Chrome/Edge/Safari do not support the `developer` manifest key.
Instead, they want to have `author` and `homepage_url` when it
comes to that metadata. This is valid for Firefox as well, so
just dropping it in.

Further add a CHANGELOG in the project, so that there is a
nice overview of what was added/changed/removed and when.